backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
ollama
python-dotenv
pillow
cairosvg
boto3
azure-identity
azure-mgmt-consumption
google-cloud-billing
requests
